Add a README.md to appease Sourcehut
Sourcehut refuses to use anything but Mistletoe for displaying README
files. This would be fine except, for some reason, it will recognise
README.* files but refuse to show them as text/plain if they are not
Markdown.

While we're here, rename the old README.adoc to syncup.1.adoc, which
better reflects its content.

Signed-off-by: Alex Minghella <a@minghella.net>